Create gprofiler dotplots

Wrapper function for the creation of a dotplot which summarises the gprofiler findings for a specified ontology.

The function reads the results for a location. It will automatically search for the 'gprofiler_results/data_files' folder in the procided location. This is should be located in the TS_results folder, but can change based on the user.

The function will attempt to create the plot with proper width and height, but these can be over-written by providing them to the function.

The function will save the plot in the location provided through the gpro_file_location parameter.

GO_dotplot_wrapper(
  gpro_file_location,
  target_ontology,
  top_n,
  custom_width = NULL,
  custom_height = NULL,
  return_plot = FALSE
)

Arguments

gpro_file_location

The location where the 'gprofiler_results' folder is

target_ontology

The targeted ontology ex: 'REAC' or 'GO:BP'

top_n

The number of top GOs to plot per cluster

custom_width

A custom value for the width of the plot

custom_height

A custom value for the height of the plot

return_plot

Boolean indicating if the plot should be returned

Value

if specified, the function will return the ggplot2 object for the dotplot