The TimeSeries class

The TimeSeries class is the main class for the timeseries pipeline It is used to store the input data along with the final processed data and any intermediate results that may be used for plot creation

Slots

exp_data

A summarized experiment which will contain the sample data and the count matrix

group_names

A vector for the group names used, should be in order for which the differential expression comparisons will occur, where the first should be the experiment and second should be the control

group_colors

A named vector indicating which colors should be associated to which group

DE_method

Either limma or DESeq2 to indicate which differential expression method is used

DE_p_filter

Character of paj or pvalue to use in the establishment of significant differential expressed genes

DE_p_thresh

A numeric value used along with the pvalue filter to establish significance. Genes will be considered significant if they are below this threshold

DE_l2fc_thresh

A numeric value for the log2FolChange threshold to establish significance where the absolute value of the genes must be greater than the threshold

Gpro_org

The organism used in gprofiler friendly format

sem_list

The list containing the semantic similarity measures for each ontology

DESeq2_obj

The normalized DESeq2 object

limma_object

The EList limma object

DE_results

A list of results for the different differential expression experiments performed

PART_l2fc_thresh

A integer indicating the log(2)foldchange threshold for genes to be PART clustered

PART_results

A list of results for the PART clustering analysis

sem_sim_org

A string indicating the annotation DBI organism to use

Gprofiler_results

A list of Gprofiler results for the PART clusters