Adds experiment data from tximport estimated reads in the form of a SummarizedExperiment

A means to add experiment data from Kallisto or Salmon to TiSA objects. The function utilizes DESeq2 to convert transcripts to genes along with a transcript to gene conversion dataframe. DESeq2 creates a dds object, the necessary elements are extracted and stored into a timeseries object instead.

This function, if used, is intended to replace the standard add_experiment_data function.

add_exp_data_tximport(
  time_object,
  sample_dta_path,
  alt_files,
  tx2gene_path,
  tx_type,
  matrix_name = "raw"
)

Arguments

time_object

A TimeSeries_Object

sample_dta_path

String which gives the csv path to the sample data

alt_files

a vector containing the path(s) to the txfile reads files. Each path must contain the associated sample name.

tx2gene_path

A path to a csv containing a dataframe showing transcript to gene conversion. The format of the dataframe will depend on the kallisto files inputed

tx_type

A string detailing from what tool the experiment comes from. For example, 'salmon', or 'kallisto'

matrix_name

Name to be give to the matrix. Default is 'raw'

Value

time_object - the updated timeSeries object