Adds experiment data in the form of a SummarizedExperiment

The function creates a SummarizedExperiment from the provided data. It reads the sample data and value files, filters them to contain only the necessary groups and then creates a SummarizedExperiment to then store this into the TimeSeris_Object

add_experiment_data(
  time_object,
  sample_dta_path,
  count_dta_path,
  limma_ID_replace = "GeneName"
)

Arguments

time_object

A TimeSeries_Object

sample_dta_path

String which gives the csv path to the sample data

count_dta_path

String which gives the path to the count or microarray data

limma_ID_replace

The name of gene id's to use from the 'genes' dataframe in the Elist

Value

count_matrix

Examples

path <- tempfile()
bfc <- BiocFileCache(path, ask = FALSE)
bfc_cache<-slot(bfc,'cache')
write_example_data_to_dir(target_dir=bfc_cache)
my_path_data<-paste0(bfc_cache,'/data/PBMC/raw_counts_TS')
my_path_sample_dta<-paste0(bfc_cache,'/data/PBMC/sample_file.csv')
graph_vect<-c("#e31a1c","#1f78b4")

TS_object <- new('TimeSeries_Object',
                 group_names=c('IgM','LPS'),group_colors=graph_vect,DE_method='DESeq2',
                 DE_p_filter='padj',DE_p_thresh=0.05,DE_l2fc_thresh=1,
                 PART_l2fc_thresh=4,sem_sim_org='org.Hs.eg.db',Gpro_org='hsapiens')

TS_object <- add_experiment_data(TS_object,sample_dta_path=my_path_sample_dta,count_dta_path=my_path_data)