Calculate MDS matrix

Function which calculates semantic distance and MDS from a dataframe of GOs

The function uses a GOSemSim object to calculates the termdistance, followed by a MDS calculation. The resulting dataframe is formatted to be used in plot_MDS

This function is used to create a GO term plot for WGCNA modules

calculate_and_format_MDS(GO_df, semantic_data, measure = "Wang")

Arguments

GO_df

The dataframe of GOs as returned by gprofiler_cluster_analysis

measure

the measuring method for semantic distance

sem_data

semantic similarity data as created by the godata function

Value

res.mds A dataframe containing the MDS dimensions, along with groupings and colors which are equivalent to the cluster colors

Examples

TS_object<-create_example_object_for_R()
#> Warning: 'data/PBMC' already exists
#> Warning: 'data/PBMC/raw_counts_TS' already exists
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#> converting counts to integer mode
#Perform conditional differential gene expression analysis
TS_object<-conditional_DE_wrapper(TS_object)
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#Extract genes for PART clustering based on defined log(2)foldChange threshold
signi_genes<-select_genes_with_l2fc(TS_object)

#Use all samples, but implement a custom order. In this case it is reversed
sample_data<-exp_sample_data(TS_object)
samps_2<-sample_data$sample[sample_data$group==TS_object@group_names[2]]
samps_1<-sample_data$sample[sample_data$group==TS_object@group_names[1]]

#Create the matrix that will be used for PART clustering
TS_object<-prep_counts_for_PART(object=TS_object,target_genes=signi_genes,scale=TRUE,target_samples=c(samps_2,samps_1))
TS_object<-compute_PART(TS_object,part_recursion=10,part_min_clust=10,dist_param="euclidean", hclust_param="average")
#> computing PART clusters
TS_object<-run_gprofiler_PART_clusters(TS_object) #Run the gprofiler analysis
#> running Gprofiler on PART clusters
#> Gprofiler for C1
#> Gprofiler for C10
#> Gprofiler for C11
#> Gprofiler for C12
#> Gprofiler for C13
#> No results to show
#> Please make sure that the organism is correct or set significant = FALSE
#> Gprofiler for C14
#> Gprofiler for C15
#> Gprofiler for C16
#> Gprofiler for C17
#> Gprofiler for C2
#> No results to show
#> Please make sure that the organism is correct or set significant = FALSE
#> Gprofiler for C3
#> Gprofiler for C4
#> Gprofiler for C5
#> Gprofiler for C6
#> Gprofiler for C7
#> Gprofiler for C8
#> Gprofiler for C9
#> No results to show
#> Please make sure that the organism is correct or set significant = FALSE
#Results saved to created directory
dir.create('example_location')
#> Warning: 'example_location' already exists
gpro_res<-gprofiler_cluster_analysis(TS_object,'GO:BP',save_path=dir.create('example_location'))
#> Warning: 'example_location' already exists
#> Error in normalizePath(dirname(file), mustWork = TRUE): path[1]="FALSE/gprofiler_results/figures": No such file or directory
GO_clusters<-gpro_res[['GO_df']]
#> Error in eval(expr, envir, enclos): object 'gpro_res' not found
TS_object <- add_semantic_similarity_data(TS_object,ont_sem_sim='BP')
#> Error in loadNamespace(name): there is no package called ‘org.Hs.eg.db’
plot_data<-calculate_and_format_MDS(GO_clusters,TS_object@sem_list)
#> Error in termSim(GO1, GO2, semData, method = measure): object 'GO_clusters' not found