compute PART clusters

The function set's a seed for reproducibility of results, it then uses the part function from the clusterGenomics package to establish which genes belong to what clusters. The clusters are then ordered using hierarchical clustering. This may result in some PART cluster being split, therefore the order is adjusted to put them together. We first order using hierarchical clustering for visual purposes.

compute_PART(
  object,
  part_recursion = 100,
  part_min_clust = 10,
  dist_param = "euclidean",
  hclust_param = "average",
  custom_seed = NULL,
  custom_matrix = NULL,
  return_as_object = TRUE
)

Arguments

object

A timeseries object

part_recursion

The number of recursions for PART calculation

part_min_clust

The minimum number of genes per cluster

dist_param

The distance parameter for clustering

hclust_param

The hierarchical clustering method/parameter to be used

custom_seed

The seed inputed (if any)

custom_matrix

Allows the input of a custom matrix instead of taking it from the object

return_as_object

Boolean indicating if the results should be returned within the submitted object or as a list

save_name

the file name to save the PART data, if NULL no saving

Value

The timeseries object with the PART results added

Examples

TS_object<-create_example_object_for_R()
#> Warning: 'data/PBMC' already exists
#> Warning: 'data/PBMC/raw_counts_TS' already exists
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#> converting counts to integer mode
#Perform conditional differential gene expression analysis
TS_object<-conditional_DE_wrapper(TS_object)
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#Extract genes for PART clustering based on defined log(2)foldChange threshold
signi_genes<-select_genes_with_l2fc(TS_object)

#Use all samples, but implement a custom order. In this case it is reversed
sample_data<-exp_sample_data(TS_object)
samps_2<-sample_data$sample[sample_data$group==TS_object@group_names[2]]
samps_1<-sample_data$sample[sample_data$group==TS_object@group_names[1]]

#Create the matrix that will be used for PART clustering
TS_object<-prep_counts_for_PART(object=TS_object,target_genes=all_genes,scale=TRUE,target_samples=c(samps_2,samps_1))
#> Error in prep_counts_for_PART(object = TS_object, target_genes = all_genes,     scale = TRUE, target_samples = c(samps_2, samps_1)): object 'all_genes' not found
TS_object<-compute_PART(TS_object,part_recursion=10,part_min_clust=10,dist_param="euclidean", hclust_param="average")
#> computing PART clusters
#> Error in array(x, c(length(x), 1L), if (!is.null(names(x))) list(names(x),     NULL) else NULL): 'data' must be of a vector type, was 'NULL'