Create data for temporal heatmap

This function is intended to be used within the custom_heatmap_wrapper

The function creates a list containing the main data matrix along with three vectors l2fc_vector: A named vector containing the L2FC value for each gene of the matrix group_vector: A vector containing the new names of the groups to represent the group they are in gene_vector: A named vector containing the gene names along with the group that they are in.

The vectors are created to account for the possibility of duplicates. For example, a gene may be significant in two timepoint comparisons, and therefore have two different values. The vectors give the genes unique names based on their grouping in order to enable them to have different values if necessary.

create_temporal_heatmap_matrix(time_object, adjust_for_missing_samples = TRUE)

Arguments

time_object

A time series object

adjust_for_missing_samples

Boolean indicating if missing samples should be compensated for (NA introduced). Currently, there is no reason to have this set to False.

Value

A list containing the main data matrix (count values) and the three vectors described above.

Examples

TS_object<-create_example_object_for_R()
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#> converting counts to integer mode
#Perform conditional differential gene expression analysis
TS_object<-temporal_DE_wrapper(TS_object)
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
heat_dta<-create_temporal_heatmap_matrix(TS_object)