Create heatmap for differential gene expression results

A wrapper function which calls the necessary functions to create a heatmap illustrating the significant DEGs of a given differential expression type (DE_type)

custom_heatmap_wrapper(
  time_object,
  DE_type,
  log_transform = TRUE,
  plot_file_name = "custom_DEG_heatmap",
  adjust_missing_temp_samples = TRUE
)

Arguments

time_object

A time object

DE_type

The type of heatmap being created (conditional or temporal)

log_transform

Boolean indicating if the results should be log transformed on the heatmap

plot_file_name

The file name to be given to the heatmap once it is saved

adjust_missing_temp_samples

Boolean indicating if missing temporal samples should be compensated for (NA introduced). Currently, there is no reason to have this set to False.

Value

none

Examples

TS_object<-create_example_object_for_R()
#> Warning: 'data/PBMC' already exists
#> Warning: 'data/PBMC/raw_counts_TS' already exists
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#> converting counts to integer mode
#Perform conditional differential gene expression analysis
TS_object<-conditional_DE_wrapper(TS_object)
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#Run function, results saved to main directory
custom_heatmap_wrapper(TS_object,DE_type = 'conditional')
#> agg_png 
#>       2