Create an MA plot

The function creates a 'four quadrant' MA plot, where the FDR and log2FoldChange thresholds dictate the significant up-regulated category, the significant downregulated category, the significant low regulation category and the non-significant category

Genes of interest are labeled in a rectangle for visibility and will have the same color as the category in which they are in.

The plot is created using ggplot2, to save a plot the ggsave() function is recommended It is also recommended to use the following parameters to save the plot. dpi=300 width=21 height=19 units='cm'

maplot_alt(
  DE_res,
  genes_of_interest = c(),
  filter_choice,
  l2FC_thresh = 1,
  p_thresh = 0.05,
  plot_title = "MA plot"
)

Arguments

DE_res

The differential expression results to be plotted

genes_of_interest

A vector containing gene names to be labelled in the plot To not label any genes, leave as default or provide an empty vector.

filter_choice

Either padj or pvalue, the choice will be used to filter for significance

l2FC_thresh

The value used as the log2FoldChange threshold

p_thresh

The value used to establish significance from adjusted pvalue

plot_title

The title to be give to the plot

Value

None

Examples

TS_object<-create_example_object_for_R()
#> Error in TS_load_example_data(TS_object, "PBMC"): unused argument ("PBMC")
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#> Error in slot(time_object, "group_names"): object 'TS_object' not found
#Perform conditional differential gene expression analysis
TS_object<-conditional_DE_wrapper(TS_object)
#> Error in slot(time_object, "group_names"): object 'TS_object' not found
DE_res<-slot(TS_object,'DE_results')$conditional$IgM_vs_LPS_TP_1$DE_raw_data
#> Error in slot(TS_object, "DE_results"): object 'TS_object' not found
ma_plot<-maplot_alt(DE_res = DE_res,filter_choice = 'padj')
#> Error in rev(DE_res): object 'DE_res' not found