merge_MDS modules

Function finds and merges GOs from separate modules

If a go appears as a part of two or more modules, the function will remove the duplicate GOs and keep a single version which has all the relevant modules as it's source ex: blue/yellow/brown if the GO was found in the blue, yellow, and brown cluster It also finds the color which best mixes the colors of the cluster

merge_duplicate_modules(the_data)

Arguments

the_data

dataframe as returned by calculate_and_format_MDS

Value

the_data The updated inputted dataframe

Examples

TS_object<-create_example_object_for_R()
#> Error in collect(., Inf): Failed to collect lazy table.
#> Caused by error in `db_collect()`:
#> ! Arguments in `...` must be used.
#> ✖ Problematic argument:
#> • ..1 = Inf
#> ℹ Did you misspell an argument name?
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'colData': object 'TS_object' not found
#Perform conditional differential gene expression analysis
TS_object<-conditional_DE_wrapper(TS_object,vignette_run=TRUE)
#> Error in eval(expr, envir, enclos): object 'TS_object' not found
TS_object<-temporal_DE_wrapper(TS_object,do_all_combinations=TRUE,vignette_run=TRUE)
#> Error in eval(expr, envir, enclos): object 'TS_object' not found
#Extract genes for PART clustering based on defined log(2)foldChange threshold
signi_genes<-select_genes_with_l2fc(TS_object)
#> Error in eval(expr, envir, enclos): object 'TS_object' not found

#Use all samples, but implement a custom order. In this case it is reversed
sample_data<-exp_sample_data(TS_object)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'colData': object 'TS_object' not found
TS_groups<-slot(TS_object,'group_names')
#> Error in eval(expr, envir, enclos): object 'TS_object' not found
samps_2<-sample_data$sample[sample_data$group==TS_groups[2]]
#> Error in eval(expr, envir, enclos): object 'sample_data' not found
samps_1<-sample_data$sample[sample_data$group==TS_groups[1]]
#> Error in eval(expr, envir, enclos): object 'sample_data' not found

#Create the matrix that will be used for PART clustering
TS_object<-prep_counts_for_PART(object=TS_object,target_genes=signi_genes,scale=TRUE,target_samples=c(samps_2,samps_1))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'table' in selecting a method for function '%in%': object 'TS_object' not found
TS_object<-compute_PART(TS_object,part_recursion=10,part_min_clust=10,dist_param="euclidean", hclust_param="average",vignette_run=TRUE)
#> Error in eval(expr, envir, enclos): object 'TS_object' not found
TS_object<-run_gprofiler_PART_clusters(TS_object,vignette_run=TRUE) #Run the gprofiler analysis
#> Error in eval(expr, envir, enclos): object 'TS_object' not found
#Set to null to not save results
gpro_res<-gprofiler_cluster_analysis(TS_object,'GO:BP',save_path=NULL)
#> Error in eval(expr, envir, enclos): object 'TS_object' not found
GO_clusters<-gpro_res[['GO_df']]
#> Error in eval(expr, envir, enclos): object 'gpro_res' not found
sem_dta<-slot(TS_object,'sem_list')
#> Error in eval(expr, envir, enclos): object 'TS_object' not found
plot_data<-calculate_and_format_MDS(GO_clusters,sem_dta)
#> Error in eval(expr, envir, enclos): object 'GO_clusters' not found
plot_data<-merge_duplicate_modules(plot_data)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'unique': object 'plot_data' not found