prep count matrix form RNAseq count files

Function requires the path to the individual count files and the samples to be included in the matrix. Files will be read based on the sample they represent the values from the different files are merged into a matrix and returned

prep_RNAseq_matrix(path_to_counts, selected_samples)

Arguments

path_to_counts

Path to the individual RNAseq count files

selected_samples

The samples to be read

Value

final_counts The formatted count matrix

Examples

write_example_data_to_dir('PBMC')
#> Warning: 'data/PBMC' already exists
#> Warning: 'data/PBMC/raw_counts_TS' already exists
my_path_data<-'data/PBMC/raw_counts_TS'
my_path_sample_dta<-'data/PBMC/sample_file.csv'

TS_object <- new('TimeSeries_Object',
                 group_names=c('IgM','LPS'),group_colors=graph_vect,DE_method='DESeq2',
                 DE_p_filter='padj',DE_p_thresh=0.05,DE_l2fc_thresh=1,
                 PART_l2fc_thresh=4,sem_sim_org='org.Hs.eg.db',Gpro_org='hsapiens')
#> Error in initialize(value, ...): object 'graph_vect' not found

TS_object <- add_experiment_data(TS_object,sample_dta_path=my_path_sample_dta,count_dta_path=my_path_data)
#> Error in slot(time_object, "group_names"): object 'TS_object' not found
groups<-slot(time_object,'group_names')
#> Error in slot(time_object, "group_names"): object 'time_object' not found
#Ensures that the order will follow the grouping order
selected_samples_1<-samp_data$sample[samp_data$group %in% groups[1]]
#> Error in eval(expr, envir, enclos): object 'samp_data' not found
selected_samples_2<-samp_data$sample[samp_data$group %in% groups[2]]
#> Error in eval(expr, envir, enclos): object 'samp_data' not found
selected_samples<-c(selected_samples_1,selected_samples_2)
#> Error in eval(expr, envir, enclos): object 'selected_samples_1' not found

#Prepare the matrix according to the differential expression method (affects input)
 final_counts<-prep_RNAseq_matrix(path_to_data,selected_samples)
#> Error in list.files(path_to_counts): object 'path_to_data' not found