Create ancestor query plots

Wrapper function which makes the necessary calls to plots ancestor curation plots

These plots find all terms which are children of the provided ancestors within the specified ontology.

The function will plot and save a dotplot as well as an interactive MDS plot

wrapper_ancestor_curation_plots(
  GO_df,
  sem_data,
  use_names = TRUE,
  target_dir = "TS_results/",
  return_plot = FALSE
)

Arguments

GO_df

The dataframe of GOs as returned by find_relation_to_ancestors

sem_data

semantic similarity data as created by the godata function

use_names

boolean indicating if names or IDs should be used

target_dir

string indicating the save location of the plots

return_plot

boolean indicating if the plot should be returned

Value

if specified, will return a list containing the ggplot2 object for the dotplot and the plotly object for the MDS plot.

Examples

TS_object<-create_example_object_for_R()
#> Warning: 'data/PBMC' already exists
#> Warning: 'data/PBMC/raw_counts_TS' already exists
TS_object <- normalize_timeSeries_with_deseq2(time_object=TS_object)
#> converting counts to integer mode
#Perform conditional differential gene expression analysis
TS_object<-conditional_DE_wrapper(TS_object)
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#> using pre-existing size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#Extract genes for PART clustering based on defined log(2)foldChange threshold
signi_genes<-select_genes_with_l2fc(TS_object)

#Use all samples, but implement a custom order. In this case it is reversed
sample_data<-exp_sample_data(TS_object)
samps_2<-sample_data$sample[sample_data$group==TS_object@group_names[2]]
samps_1<-sample_data$sample[sample_data$group==TS_object@group_names[1]]

#Create the matrix that will be used for PART clustering
TS_object<-prep_counts_for_PART(object=TS_object,target_genes=signi_genes,scale=TRUE,target_samples=c(samps_2,samps_1))
TS_object<-compute_PART(TS_object,part_recursion=10,part_min_clust=10,dist_param="euclidean", hclust_param="average")
#> computing PART clusters
TS_object<-run_gprofiler_PART_clusters(TS_object) #Run the gprofiler analysis
#> running Gprofiler on PART clusters
#> Gprofiler for C1
#> Gprofiler for C10
#> Gprofiler for C11
#> Gprofiler for C12
#> Gprofiler for C13
#> Gprofiler for C2
#> Gprofiler for C3
#> Gprofiler for C4
#> Gprofiler for C5
#> Gprofiler for C6
#> Gprofiler for C7
#> Gprofiler for C8
#> No results to show
#> Please make sure that the organism is correct or set significant = FALSE
#> Gprofiler for C9
#Create directory
dir.create('example_location')
#> Warning: 'example_location' already exists
gpro_res<-gprofiler_cluster_analysis(TS_object,'GO:BP',save_path='example_location')
GO_clusters<-gpro_res[['GO_df']]
TS_object <- add_semantic_similarity_data(TS_object,ont_sem_sim='BP')
#> preparing gene to GO mapping data...
#> preparing IC data...
target_ancestors<-c('GO:0002253','GO:0019882','GO:0002404','GO:0002339','GO:0042386',
                    'GO:0035172','GO:0002252','GO:0006955','GO:0002520','GO:0090713',
                    'GO:0045321','GO:0001776','GO:0050900','GO:0031294','GO:0002262',
                    'GO:0002683','GO:0002684','GO:0002440','GO:0002682','GO:0002200',
                    'GO:0045058','GO:0002507')
GOs_ancestors_clust<-find_relation_to_ancestors(target_ancestors,GO_clusters,ontology = 'BP')
#Plots will be returned, but also saved to directory
ancestor_plots<-wrapper_ancestor_curation_plots(GOs_ancestors_clust,TS_object@sem_list,return_plot=TRUE,target_dir='example_location/')
#> Warning: 'example_location/ancestor_plots' already exists